refactor(tsdown-driver): split artifact manifest indexing boundaries
