+++ packages/tsdown-driver/src/artifact-indexer/internal/index.ts
@@ -1,0 +1,8 @@
+export { indexArtifacts } from "../indexing.js";
+export {
+  isBundleFile,
+  isTypeFile,
+  normalizeChunkFile,
+  normalizeTsconfigPath,
+  uniqueSorted,
+} from "../validation.js";
